requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-exporter-prometheus==0.46b0
prometheus-client==0.19.0
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
